chore(types): Added types for plugin options and removed ts-ignore
