+++ packages/core/server/admin-api/bootstrap.ts
@@ -1,7 +1,7 @@
 
 
 import { getPluginService } from '../util/getPluginService';
-import { IStrapi } from '../types/strapi';
+import { IStrapi } from '../../types/strapi';
 import migrateToNativeRelation from './migrations/to-native-relation';
 import migratePluginOptionsRename from './migrations/plugin-options-rename';
